DDP-8493 I moved the file download methods to `participant-page.component.ts` and just emit an event from `file-download.component.ts`. So the feedback is in the participant-page now
